+++ src/tbot.c
@@ -24,7 +24,9 @@ static const struct mrb_data_type mrb_gem_data_type = {
 void mrb_mruby_tbot_gem_init(mrb_state *mrb)
 {
   struct RClass *tbot;
+  struct RClass *sender;
   tbot = mrb_define_module(mrb, "TBot");
+  sender = mrb_define_module(mrb, "Sender");
   DONE;
 }
